refactor(test): assert_pageserver_backups_equal
